requirements.txt
mysql-connector-python
inflection
pytest
pytest-mock
pytest-cov
coverage
cryptography
